chore: move get address to its own module
